requirements.txt
fastapi==0.115.12
pyjwt==2.10.1
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.6
supabase==2.13.0
uvicorn==0.34.0
dspy==2.6.16
openai==1.61.0
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==11.1.0
python-magic
python-docx==1.1.2
numpy==2.2.3
matplotlib==3.10.1
anthropic
psutil
PyPDF2==3.0.1
google-genai==0.6.0
